Moved logic to viewmodels. Fixed bugs, removed nesting in the design, add responsiveness.
