+++ app/build.gradle
@@ -40,6 +40,8 @@ dependencies {
     implementation 'com.github.bumptech.glide:glide:4.12.0'
     implementation 'com.google.firebase:firebase-auth:21.0.1'
     annotationProcessor 'com.github.bumptech.glide:compiler:4.12.0'
+    implementation 'com.jakewharton:butterknife:10.2.3'
+    kapt 'com.jakewharton:butterknife-compiler:10.2.3'
     implementation 'com.firebaseui:firebase-ui-auth:6.4.0'
     implementation 'com.google.firebase:firebase-bom:28.0.1'
     implementation 'com.google.firebase:firebase-storage:20.0.0'
