+++ app/src/main/java/android_courses/newsapp/SignUpViewModel.kt
@@ -2,50 +2,35 @@ package android_courses.newsapp
 
 import android.content.Context
 import android.content.SharedPreferences
-import android.os.Bundle
 import android.util.Patterns
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.widget.*
-import android_courses.newsapp.Utill.Constants.Companion.APP_PREFERENCES
+import android_courses.newsapp.Utill.Constants
 import android_courses.newsapp.Utill.Constants.Companion.IS_REGISTERED
 import android_courses.newsapp.base.BaseActivity
-import androidx.fragment.app.Fragment
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseAuthUserCollisionException
 
-class SignUpFragment : Fragment(), View.OnClickListener {
-    lateinit var progressBar: ProgressBar
-    lateinit var editTextEmail: EditText
-    lateinit var editTextPassword:EditText
-    lateinit var buttonSignUp: Button
-    lateinit var textViewLogin: TextView
-    lateinit var mSettings: SharedPreferences
-    private lateinit var mAuth: FirebaseAuth
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        mAuth = FirebaseAuth.getInstance()
-        mSettings = requireContext().getSharedPreferences(APP_PREFERENCES, Context.MODE_PRIVATE)
+class SignUpViewModel(private val context: Context) : View.OnClickListener{
+    private val mSettings: SharedPreferences by lazy {
+        context.getSharedPreferences(Constants.APP_PREFERENCES, Context.MODE_PRIVATE)
     }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_sign_up, container, false)
+    private val mAuth: FirebaseAuth by lazy {
+        FirebaseAuth.getInstance()
     }
+    private lateinit var progressBar: ProgressBar
+    private lateinit var editTextEmail: EditText
+    private lateinit var editTextPassword: EditText
+    private lateinit var buttonSignUp: Button
+    private lateinit var textViewLogin: TextView
 
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
+    fun onViewCreated(view: View) {
         editTextEmail = view.findViewById(R.id.editTextEmail)
         editTextPassword = view.findViewById(R.id.editTextPassword)
         progressBar = view.findViewById(R.id.progressbar)
         buttonSignUp = view.findViewById(R.id.buttonSignUp)
-        buttonSignUp.setOnClickListener(this)
         textViewLogin = view.findViewById(R.id.textViewLogin)
+        buttonSignUp.setOnClickListener(this)
         textViewLogin.setOnClickListener(this)
     }
 
@@ -81,19 +66,19 @@ class SignUpFragment : Fragment(), View.OnClickListener {
                         putBoolean(IS_REGISTERED, true)
                         apply()
                     }
-                    (requireActivity() as BaseActivity).fragmentRouter.openNewsFragment()
+                    (context as BaseActivity).fragmentRouter.openNewsFragment()
                 } else {
                     if (task.exception is FirebaseAuthUserCollisionException) {
                         Toast.makeText(
-                            context,
-                            "You are already registered",
-                            Toast.LENGTH_SHORT
+                                context,
+                                "You are already registered",
+                                Toast.LENGTH_SHORT
                         ).show()
                     } else {
                         Toast.makeText(
-                            context,
-                            task.exception!!.message,
-                            Toast.LENGTH_SHORT
+                                context,
+                                task.exception!!.message,
+                                Toast.LENGTH_SHORT
                         ).show()
                     }
                 }
@@ -104,7 +89,7 @@ class SignUpFragment : Fragment(), View.OnClickListener {
         when (view.id) {
             R.id.buttonSignUp -> registerUser()
             R.id.textViewLogin -> {
-                requireActivity().supportFragmentManager.popBackStack()
+                (context as BaseActivity).supportFragmentManager.popBackStack()
             }
         }
     }
