+++ app/src/main/java/android_courses/newsapp/fragments/FragmentSetting.kt
@@ -1,16 +1,29 @@
 package android_courses.newsapp.fragments
 
+import android.content.Context
+import android.content.SharedPreferences
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageButton
 import android_courses.newsapp.R
+import android_courses.newsapp.Utill.Constants.Companion.APP_PREFERENCES
+import android_courses.newsapp.Utill.Constants.Companion.IS_REGISTERED
+import android_courses.newsapp.base.BaseActivity
+import androidx.appcompat.widget.AppCompatButton
+import androidx.appcompat.widget.AppCompatImageButton
+import com.google.firebase.auth.FirebaseAuth
 
 class FragmentSetting : Fragment() {
-
-    lateinit var backButton: ImageButton
+    private val mAuth: FirebaseAuth by lazy {
+        FirebaseAuth.getInstance()
+    }
+    private val mSettings: SharedPreferences by lazy {
+        requireContext().getSharedPreferences(APP_PREFERENCES, Context.MODE_PRIVATE)
+    }
+    private lateinit var backButton: AppCompatImageButton
+    private lateinit var buttonLogout: AppCompatButton
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
@@ -21,10 +34,18 @@ class FragmentSetting : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        buttonLogout = view.findViewById(R.id.button_log_out)
         backButton = view.findViewById(R.id.imageButton2)
+        buttonLogout.setOnClickListener {
+            mAuth.signOut()
+            with (mSettings.edit()) {
+                putBoolean(IS_REGISTERED, false)
+                apply()
+            }
+            (requireActivity() as BaseActivity).fragmentRouter.openSplashFragment()
+        }
         backButton.setOnClickListener {
             requireActivity().supportFragmentManager.popBackStack()
         }
     }
 }
-
