+++ app/src/main/java/android_courses/newsapp/fragments/NewsFragment.kt
@@ -1,15 +1,11 @@
 package android_courses.newsapp.fragments
 
-import android.content.Context
-import android.content.SharedPreferences
 import android.os.Bundle
 import android.util.Log
 import android.view.View
 import android_courses.newsapp.NewsViewModel
 import android_courses.newsapp.NewsViewModelProviderFactory
 import android_courses.newsapp.R
-import android_courses.newsapp.Utill.Constants.Companion.APP_PREFERENCES
-import android_courses.newsapp.Utill.Constants.Companion.IS_REGISTERED
 import android_courses.newsapp.Utill.isVisible
 import android_courses.newsapp.adapter.NewsAdapter
 import android_courses.newsapp.base.BaseActivity
@@ -19,26 +15,16 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.google.firebase.auth.FirebaseAuth
 import kotlinx.android.synthetic.main.fragment_news.*
 
 class NewsFragment : Fragment(R.layout.fragment_news) {
-    lateinit var mAuth: FirebaseAuth
-    lateinit var mSettings: SharedPreferences
-    lateinit var buttonLogout: AppCompatImageButton
-    lateinit var buttonSelection: AppCompatImageButton
-    lateinit var buttonSettings: AppCompatImageButton
-    lateinit var viewModel: NewsViewModel
+    private lateinit var buttonSelection: AppCompatImageButton
+    private lateinit var buttonSettings: AppCompatImageButton
+    private lateinit var viewModel: NewsViewModel
     private val newsAdapter: NewsAdapter by lazy {
         NewsAdapter {
             Log.d("ClickOnArticle",
                     it.description) } }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        mAuth = FirebaseAuth.getInstance();
-        mSettings = requireContext().getSharedPreferences(APP_PREFERENCES, Context.MODE_PRIVATE)
-    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -59,22 +45,11 @@ class NewsFragment : Fragment(R.layout.fragment_news) {
             pagination_progress_bar.isVisible(visibility)
         })
 
-        buttonLogout = view.findViewById(R.id.log_out_btn)
-        buttonLogout.setOnClickListener {
-            mAuth.signOut()
-            with (mSettings.edit()) {
-                putBoolean(IS_REGISTERED, false)
-                apply()
-            }
-            (requireActivity() as BaseActivity).fragmentRouter.openSplashFragment()
-        }
-
         buttonSelection = view.findViewById(R.id.img_settingsLine)
+        buttonSettings = view.findViewById(R.id.img_settings)
         buttonSelection.setOnClickListener {
             (requireActivity() as BaseActivity).fragmentRouter.openSelectionFragment()
         }
-
-        buttonSettings = view.findViewById(R.id.img_settings)
         buttonSettings.setOnClickListener {
             (requireActivity() as BaseActivity).fragmentRouter.openSettingsFragment()
         }
