+++ app/src/main/java/android_courses/newsapp/fragments/SignUpFragment.kt
@@ -1,0 +1,29 @@
+package android_courses.newsapp.fragments
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android_courses.newsapp.R
+import android_courses.newsapp.SignUpViewModel
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+
+class SignUpFragment : Fragment() {
+    private val viewModel: SignUpViewModel by lazy {
+        SignUpViewModel(requireActivity())
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_sign_up, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        viewModel.onViewCreated(view)
+    }
+}
