+++ app/src/main/res/layout/fragment_news.xml
@@ -9,21 +9,6 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     tools:context=".fragments.NewsFragment">
 
-    <ImageButton
-        android:id="@+id/log_out_btn"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentEnd="false"
-        android:layout_marginTop="10dp"
-        android:layout_marginBottom="5dp"
-        android:src="@drawable/ic_logout"
-        app:layout_constraintBottom_toTopOf="@id/rv_breaking_news"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.742"
-        app:layout_constraintStart_toEndOf="@+id/img_settingsLine"
-        app:layout_constraintTop_toTopOf="@id/constraint"
-        app:layout_constraintVertical_bias="0.0" />
-
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_breaking_news"
         android:layout_width="match_parent"
@@ -64,9 +49,11 @@
     <ImageButton
         android:id="@+id/img_settings"
         android:layout_width="wrap_content"
-        android:layout_height="0dp"
+        android:layout_height="wrap_content"
         android:layout_alignParentEnd="false"
         android:background="@android:color/transparent"
+        android:layout_marginTop="10dp"
+        android:layout_marginBottom="5dp"
         android:src="@drawable/ic_component_17"
         app:layout_constraintBottom_toTopOf="@id/rv_breaking_news"
         app:layout_constraintEnd_toEndOf="parent"
